requirements.txt
Flask==3.0.3
gunicorn==21.2.0
cloudinary==1.41.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3